fix: store restructre, TS added to RTK
