--- src/redux/store.tsx
@@ -1,8 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import userReducer from './userSlice';
-
-export const store = configureStore({
-  reducer: {
-    user: userReducer,
-  },
-});
