+++ src/store/user/actions.ts
@@ -1,0 +1,4 @@
+import { UserState } from '..';
+import { createAction } from '@reduxjs/toolkit';
+
+export const setUser = createAction<UserState>('user/setUser');
