+++ src/store/user/selectors.ts
@@ -1,0 +1,3 @@
+import { UserState } from '..';
+
+export const selectUser = (state: { user: UserState }) => state.user.user;
