+++ src/store/user/slice.tsx
@@ -1,0 +1,16 @@
+import { UserState } from '..';
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialState: UserState = {
+  user: {
+    email: '',
+    token: '',
+  },
+};
+export const userSlice = createSlice({
+  name: 'user',
+  initialState,
+  reducers: {},
+});
+
+export default userSlice.reducer;
